Remove obsolete config options in libvirt plugin
The libvirt detection plugin has been using the 'sql_connection'
configuration value from nova.conf, but while that is still present in
DevStack, it has been deprecated in OpenStack as of Havana.  Also, it
is completely unused, a relic of an earlier version of the detection
plugin.

Furthermore, the 'identity_uri' value only seems to exist in the
DevStack nova.conf, and is not present in OpenStack.  This value is now
built using auth_* configuration values, which exist in both.

Change-Id: I53269c16773013df213f2acbe02c0e1028d7e81b