Add the ability to amplify metric payloads
It may be helpful, for load testing, for one node to be able to send
extra metrics, multiples of its typical payload.  One node could thus
impersonate two or three or ten in terms of the total number of metrics
submitted.

This patch adds this ability through use of an 'amplifier'
configuration option.  If set above zero, each measurement will be
duplicated, with an additional 'amplifier' dimension, making it a
unique metric.  The higher 'amplifier' configuration value is raised,
the more multiples of each metric will be submitted.

To configure, run monasca-setup with the --amplifier parameter, as in
'--amplifier 2' for two additional sets of metrics per API payload.

Change-Id: Ibc277bb31bca449510fc7bb6d5520b329a78aea4