Fixed issue with agent startup if Keystone is down
The agent was not starting properly if Keystone was unavailable and
no error message was output. Moved creation of Keystone object from
the AgentCheck class to the HTTPCheck class in the collector.  Also,
cache the token in the Keystone class to keep from hitting the Keystone
API multiple times.  If Keystone is unavailable, cache the messages.

Change-Id: I4f0fd39ac3ac32b4a4b219e064c0f12b6286ee87