Remove functionality that sends 0 values when counters don't report
Also I removed the expiry check on flushing metrics, there is already a
check on submission to see if a submitted metric is old. This would only
catch metrics that were submitted correctly but took to long to be
processed by the collector.

Lastly I removed the MetricsBucketAggregator and combined all things
into a single aggregator. Only statsd was using the bucket aggregator
and there is no advantage over the other.

Change-Id: I1a2a0c9941a95558036233e61ac3f5f395ed54d5