Adding value_meta to certain metrics
The value_meta parameter was recently added to the
monasca-api.  This allows the addition of meta-information
to be sent with any metrics that are sent from the agent.
value_meta was added for the http_status, nagios_wrapper,
and host_alive checks in the agent.  Also, consolidated
some of the duplicate exception handling and reformatted
some code blocks.

Change-Id: I3d9e9a8e54cd2582bde47c7c128a3788b821c156