Re-sync Agent Collector's virt dir with Ceilometer
The code in monasca_agent/collector/virt is derived from Ceilometer
code, but has fallen out of date and should be re-merged.

The latest Ceilometer code, and therefore this patch, adds libvirt
support for VM resident memory usage inspection, as well as disk
capacity and utilization.  However, these have yet to be thoroughly
vetted or added to the Agent libvirt plugin as additional VM metrics.

Change-Id: Ia30a6059fe4e54b04cb37851f270a9f57d86ae19