Refactors swift metrics into separate check
This patch separates out swift metrics into separate checks. Earlier
all the swift metrics which were wrapped around icinga were part
of service plugin. With this, the status and messages of each diagnostic
was part of value_meta. This causes shortcoming since alarm definition
cannot be based on value_meta.

With this change, all the swift metrics are now part of separate check,
which each of values gauged separately. The services are still part of
detection plugin.

Change-Id: I53934858c07450bce33290e48adfde65949bd0c4