Check for local crash dumps
Add a check that monitors crash dumps. The specified crash directory
is scanned for sub-directories named 'YYYYmmddHHMM' which contain the
dump files. The 'dump_count' metric is the number of sub-directories
(i.e., the number of crash dumps present) and the 'value_meta' dict
contains the date-/timestamp of the most recent crash.

Implements: blueprint check-for-crashes

Change-Id: I20b584c68644ff2e76baabab78e965e682759aa0