Ensure libvirt's hostnames are used by collector
A recent change to MetricsAggregator (195214) broke hostnames when
passed as parameters to gauge().  This affected libvirt client metrics,
which alter the hostname in this manner.  To set the hostname in a way
compatible with 195214, I now delete the 'hostname' dimension for client
metrics, which was unused, anyway.

In addition, I moved the setting of dimensions to above the first
gauge() call so that all metrics are submitted with the correct
dimensions.  To prevent a regression in change 199323, I've wrapped
the dimension setup within a try/except block.

Change-Id: If3dfdc828613a45ddb78dc4653779a91b9240bdd