Add reasonable default timeouts for host_alive
The host_alive plugin supports timeouts for ping and SSH connection
tests, as well as specifying the port number to be used for SSH
tests.  However, when host_alive is configured through monasca-setup
(or the monasca_agent_plugin Ansible module, for that matter), there
are no defaults set, leaving it up to the system, and thus creating
unexpected and/or undesirable results.

This changes adds some reasonable default timeout values to the
detection plugin, along with the default SSH port.  There are also
some minor PEP8-compliance changes.

Change-Id: I8bf8cf8f31ac28a18f316b49d40b8cac49814f05