Add process monitoring for LBaaS agents
Add neutron-lbaas-agent (LBaaS V1) and neutron-lbaasv2-agent (LBaaS
V2) to the neutron detection plugin.  Because the string
"neutron-lbaas-agent" can be both a process name and log file name,
the process monitor is susceptible to false positive matching on that
string. Use a longer part of the python path to disambiguate this

Change-Id: I3081639a6f36a276bab2f9eb1b9b39a5bef452f1