Limit libvirt Nova calls to the local compute node
On OpenStack deployments with an excess of 1,000 provisioned
VMs, the libvirt plugin may not publish metrics for some or
all of the hosted VMs on each compute node.  This is due to
the default osapi_max_limit of 1000, limiting the number of
results from a Nova API query.

Instead of loading instance data for all Nova VMs, this change
limits the query to just VMs hosted on the local compute node.  It
assumes the hostname in agent.yaml matches the hostname in used by
Nova in OS-EXT-SRV-ATTR:host.

Change-Id: Id18abebc550f5bffed553248088f56d5b8a9bb63