Pass in all credentials to monclient for monclient token renewal
The monascaclient HTTPClient will try one time to get a new
keystone token when the token expires.  This works well if
it has all the credentials upon initialization.  Now the agent
passes in all needed keystone credentials so that it should
not normally see the unauthorized exception.  The HTTPClient
will try one time to get the token if that fails it will send
the Unauthorized exception and the agent can try one more time.

Change-Id: Ib43a6222cfc4b973fa5ee0809c0e6d2c3b7cd72d