libvirt detection plugin adds bad instances
Without this change, running the detection plugin results in
instances being added to /etc/monasca/agent/conf.d/libvirt.yaml that
only contained a built_by value of Libvirt. One instance would be created
for each run of the detection plugin. For example, after two runs of the
detection plugin, this would be at the end of
/etc/monasca/agent/conf.d/libvirt.yaml:

instances:
- built_by: Libvirt
- built_by: Libvirt

This doesn't actually cause any real problems with running the
libvirt detection plugin except the agent is restarted every time
the plugin is run even if there are no changes. That is because the
code thinks a new instance is being added so it has to restart the
agent.

Change-Id: Ic1a5a87b4c1c184d1e971edbb56c060dbd2473d7