Allow config file 'instances' to be empty
This change allows slightly more lenient processing of Agent config
files, where 'instances' is still required in the .yaml file, but may
be an empty list.

An example of where this is necessary is the libvirt plugin, where
it is unreasonable to have all VMs enumerated as 'instances' in the
.yaml file, but are instead provided by the hypervisor.

The libvirt plugin itself needed a minor modification to work with
empty instances.  These changes are necessitated by the monasca-setup
change https://review.openstack.org/239815

Change-Id: Idc67df9e5d9195f65fe16b9d67478e1de8e61eb3