Allow overriding values using args in ServicePlugin
Some of the subclasses of ServicePlugin allow overriding the
service_api_url using args. This could be useful for all
plugins so move that code into ServicePlugin.

Also allow overriding of process_names, search_pattern,
overwrite and template_dir.

Remove the code in the ironic and trove plugins that duplicated
this code.

Added a unit test for this code.

Change-Id: Ibd97effa6d5bff0cf621e4bdcbe6f5b781d195f4