Replace host_alive test with hypervisor VM states
Currently, the host_alive_status metric is a combination of ping tests
and general VM up/down state from the hypervisor isActive().  If ping
tests are not supported in the environment or by the VM's security rules,
host_alive_status becomes practically useless.

This change proposes to make host_alive_status a reliable, stable metric
consisting of one of eight possible status codes, as reported by the
hypervisor, including shut off/paused/suspended, etc.  A value_meta
'detail' tag includes a human-readable translation of the status code,
and the codes are also added to the plugin documentation.

In addition, this change renames the 'ping_only' configuration parameter
to 'alive_only' as a means to gauge the general status of all VMs without
the additional load of performance metrics, for situations where the latter
is overwhelming.  If alive_only is True, only the VM status metrics and
aggregate metrics are published.

The 'ping_status' metric serves as the new name for ping-based testing of
VMs.

Change-Id: I77c87b222de0aa9882d1761e8cade87ef1719d7d