Speed up libvirt cache writes, fix corruption
This patch changes the libvirt cache format from YAML to JSON.  Writing
out the caches in YAML on hosts with a large number of VMs can be slow,
~760ms for 250 VMs, and switching to JSON reduces this write time to
~32ms for the same number of VMs.

The change also removes the _is_cache_corrupt function, replacing it
with improved exception handling.  Testing shows that this change
captures and rebuilds missing, zero-byte, and truncated cache files,
adding a WARNING to the log file if/when either cache is rebuilt.

Change-Id: Ie57dcf4484d4d2d4d8c35a0978275b9c1475db97