Include existing system PATH in nagios_wrapper
Certain Nagios checks may be written in a manner that assumes common
executable-containing directories (like /bin or /usr/bin, for example)
to exist in $PATH, but in nagios_wrapper, the PATH is blanked out, except
when replaced with the path(s) specified in the init_config's check_path.
For most/all standard Nagios checks, this is fine, but for many others,
assuming a PATH that includes the likes of /bin /usr/bin, /usr/local/bin,
etc. would cause the check to fail.

One solution may be to require all Nagios checks to specify the full
path to the executables they call, but the more accommodating fix would
be include the system PATH as well as check_path, as proposed in this
patch.

Change-Id: I943c58a579bb3440c41f8b5059c2128262d24bdf