Improve output of host_alive plugin's ping tests
If a ping test fails, it returns a status code != 0, which the host_alive
plugin currently uses for a catch-all "failed to execute" message in the
value_meta 'error' field.  This is wordy and confusing for ping failures.

Additionally, if the ping command fails for an OS-level problem (missing
or no longer executable), there is no handler for that exception.

This patch resolves these two issues, removes some redundant logging, and
changed the log level of a normal ping failure.

Change-Id: I038ffcdc8577bf910e8229ff06f67bcb46cac47b