Switch Agent thread count test to global threshold
Currently, the multi-threaded ServicesCheck class will compare the
thread number allocated to each plugin against the total number of
threads in use by the Collector.  If several ServicesCheck-based plugins
are in active use, though, a normal and reaonsable number of Collector
threads would throw an exception due to an "exploding" thread count.

This change replaces the existing apples-to-oranges thread count
comparison with one global maximum number of threads.  In the highly
unlikely event of runaway thread allocation, this would provide a level
of protection, while still allowing normal operation of a large numbe
of multithreaded plugins.

Change-Id: I263fb25fab71805a24f45e5455c79c4c2d8c81e3