Remove vendor-specific paths on LBaaS agent executables
The monasca agent uses simple string matching in the process detection
plugin.  This can result in false positive matches in the case where a
process name appears elsehwere in the "ps" output.  In particular,
"neutron-lbaas-agent" is used for both process and log file names.

The initial attempt at solving this used full path names for the LBaaS
agents, but these paths were only valid for specific vendor distributions.
This patch reverts to using only "neutron-lbaasv2-agent" for the V2
LBaaS agent, and uses "bin/neutron-lbaas-agent" for the V1 agent.

While it is possible that a vendor would deliver the neutron-lbaas-agent
in a directory other than "bin", they would be diverging from the
layout generally assumed by the overall neutron project.

Change-Id: I6654d99ea060e26b7cdd5b0bdb1a1c0a2537ba3b