Libvirt updates: ping check fix, CPU time metric
This patch corrects a function of the libvirt ping check where security
group rules lacking a remote_ip_prefix cause the plugin to fail.

Also, a new metric has been added, vm.cpu.time_ms, which reports the
cumulative CPU time (in milliseconds) for VMs.  It is a metric designed
just for Operations, useful for metering and billing, and not exposed
to the VM owner.  There's a corresponding documentation update as well.

Change-Id: If1319fcb0033c9dbce4ddc5f87831bc60213ded8