Use setcap instead of sudo in libvirt VM ping test
It may be undesirable for administrators to allow password-less
sudo access to /sbin/ip for the mon-agent user.  This change replaces
the sudo requirement with a self-contained setcap solution.  The
monasca-setup detection plugin will create a local copy of /sbin/ip,
lock down the permissions, and add the cap_sys_admin capability to it
so that the mon-agent user can execute an in-namespace ping without
the need for a risky NOPASSWD sudoers.d entry.

This patch also fixes a minor issue with possible uncaught exceptions
during nova process detection, bringing psutil exception handling
current.

Change-Id: Ic75107a5ec0df42d732592173b874f0835e4ca98