Validate valueMeta for measurement
If a plugin tries to give invalid valueMeta for a Measurement, the
API will reject it. Since the agent batches measurements together,
this will cause rejection of the whole batch so it is better to
just reject these measurements before sending them. It also makes
it easier to determine which plugin is sending the invalid measurements

Change-Id: Id9d14de5e7ad6a11d0c8b284a3cd59bd9ecfafbd