Change log level for permission denied on certain process checks.
Right now we log a warning and we do not believe it should be that
high. Also fixed up the documentation for this and removing some
process checks we do not gather anymore

When we do get permission denied when trying to grab metrics we
now do not report any measurement. We used to report 0.

Change-Id: I64102f618ce2293590c6b516b050643a73fdd382