A couple small fixes to the postfix plugin
The postfix plugin is failing due to InvalidValue: "invalid value 0
is not of number type for metric postfix.queue_size" since it's
returning a str value instead of an int, so this patch resolves that.

Also, the detection plugin was hard-coded to look for the mon-agent
user, so this is changed to detect the user based off of the owner
of the agent configuration file.

Change-Id: Iaca266493c1fba3ca9ce8747f0152b7b56838573