Allow Storm 1.0.x to be used as well as 0.9.x
In 1.0.x, the class name start with org.apache instead of backtype.
So, I removed the backtype part so either class names that start
org.apache or backtype will be found.

To ensure it only find real storm processes, I also changed it to
look for storm.daemon.supervisor and storm.daemon.nimbus on the initial
check

Change-Id: I367dd319b77979c44ccc0acd89503aabc3ec114b