[monasca-agent] Constraint tox targets with upper-constraints
This review is to prepare for adding monasca-agent into openstack
global-requirements. After this project gets into global requirements,
it will be in upper-constraints.txt. If we make changes on this
specific project, we need to replace it in the upper-constraints with
source based specifier before applying it to pip install.

Change-Id: Id4d6dc8648285c9131f7b4626524742f271e9f26