Refactor process test
The tests relied on psutils Python module which is not part of monasca.
The used functions and classes are substituted by mock objects that
simulate their behavior, making the test independent of third party
modules.

Changed the assertTrue calls to more specific asserts.

common.py is refactored to make it pep8 compatible

Change-Id: I6badfd3612ba52f55b73219b4581e94ff79cc73b