TLS mysql plugin support
New SSL optional arguments to support TLS mysql
connections.  Arguments can be passed or a
default config file can be used to pass
the connection arguments.  When host is
localhost, unix_socket is needed by the
pymysql connect call.  Host cannot be
localhost when using SSL options for
the connection.

Change-Id: Ia600a7df1923d379eec341e604cc77f7378bf3b9