port_id dimension for port metrics causes plugin to fail
For VM port metrics, Neutron port attribute such as port ID is
obtained only in few cases. The availability of port cache for
a VM should not be assumed.
This patch checks if port cache for a VM is available before
using it.

Few cases when port cache is empty or not built are:
(a) When neutron api failed to return any response.
(b) When ping_check parameter is set to False in config.
(d) When there are no Neutron ports.

Change-Id: I9b9bfc769a172b301c78dc142ea4697e2bd63b89
Closes-Bug:1591245