Remove measurement transformation from forwarder
The forwarder modified the timestamp and possibly the tenant_id of every
measurement sent to it.  This modification does the processing on those
attributes in the collector and sends the forwarder the final set of data.

Value meta can't be part of the context used to determine if we have a new
metric.  Value meta can change while the metric itself remains the same which
means that we construct a new metric object each time we get a new value meta
value.  This doesn't have any effect on Gauge metrics but Counter and Rate
metrics return invalid data as value meta changes.  To fix this value meta is
now applied to each sample instead of being assigned during __init__.

Change-Id: I1d6a52266755b2e20149ef7ec6eb76fe2459b2a8
Closes-Bug: #1596686