Modified rate metrics calculation
For VM ports and the disk metrics, actual Rate metric that is
calculated currently is current - previous value. It was missed to divide the collection interval.
This patch fixed the code by dividing the collection interval to the value_diff.

time_diff is calculated by currenttime - cache_time from the cache file.

Change-Id: I8b7ad88d9ef95159bae5f155d24a4f144f7de2e0
Closes-Bug: 1595816