apache use_server_status_metrics to filter
We had a request to only setup the apache process metrics.
The new argument use_server_status_metrics defaults to True.
If this argument is passed in as false, we will not setup
the server-status url metrics.

I also modified the detection plugin to raise exceptions
after an error is logged to be consistent with our
other detection plugin behavior.
Raising exceptions makes them obvious when calling
monasca-setup during deployment.

Change-Id: I3ad4d93cdc15b83c1bf155a9807f5b83500a75ba