Eliminate use of hard-coded persister admin type/port
The http healthcheck for the Monasca persister was configured using
hard-coded type/port values of http/8091 instead of using the values
specified in the yaml configuration file. This change replaces the
hard-coded values with what is specified in the configuration file
if available, otherwise default values of http/8091 are used.

Change-Id: Ie29b7a12dfd7b71868a250ee906c98b92decf26e