Add Freezer API and Scheduler detection plugins
This adds plugins basic service plugins to detect Freezer's API and
Scheduler. While the scheduler only supports basic process monitoring,
the API also makes use of HTTP status checks.

Detection for freezer-api requires that mod_wsgi is configured to
include 'freezer-api' in the process name, and that the API is
accessible at 127.0.0.1 for health checks.

Change-Id: I9ea7963fb8788fdb6e11e85fccb793bb14530d16