For detection, treat libvirt vm_probation as int
Previously, when the libvirt detection plugin was called, vm_probation
would be written out as a string which would cause
the check plugin to fail. Now, vm_probation is converted to an
integer. If the conversion fails, a message is output and vm_probation
is set its default. This is consistent with the handling of the
other parameters to the libvirt detection plugin.

Fix nova_refresh as well

Change-Id: I77432a7200e2184d574ce125858853514d95a840
Closes-Bug: #1611905