Apache detection
Change method for find process from find_process_cmdline to
find_process_name. If in system is process which name contains
httpd (e.g. static_httpd) but it is not Apache, plugin should not
try to configure detection for it.
Add to detection checking if configuration file exists or
args were given.

Change-Id: I6f54cbea7c4fa2ce587f43d6808e97b3d644ab6e