Fixes WMI related checkers
Adds computer argument to the WMI constructor used in the
iis and wmi_check checkers. This way, they are compatible
with both WMI and PyMI (which is a faster, better alternative
to WMI).
Fixes bug in the wmi_check, where a private method was called with
fewer arguments that it should.

Partially-Implements: blueprint add-windows-support

Change-Id: I54688b096b750a46fdefae5918ec23b47987413e