Add service_type, endpoint_type & region_name querying Keystone Catalog
Monasca URL is pulled from Keystone Service Catalog when not given
explicitly in agent.yaml, but only the URL with service_type=monitoring
and endpoint_type=publicURL will be retrieved.

These two options together with the region_name have been included in
the agent.yaml template and may be configured by monasca-setup, so that
the KSClient object is initialized with such keywords. This client will
retrieve the URL from Catalog accordingly.

Change-Id: Icfbfa7c46db648e3d3c27f8e5c72d02957c778f1
Closes-Bug: #1625573