Expand libvirt and ovs plugins to publish entity names
This is useful in dashboards where we display libvirt and ovs
metrics for a physical host, it will elimnate the step of translating
vm and tenant ids if needing to investigate project usage.

The libvirt plugin metadata config array will now support
'vm_name' and 'tenant_name', and the ovs plugin will now
support the metadata config value for 'tenant_name'.

Change-Id: If76418b83119a12b8534fb7f7cb224339f357be8