Add a target_hostname to Host Alive plugin
To handle the case where the checked system has multiple IP Addresses
and the network name to be used for liveness checking is not the same as
the usual name used to identify the server in Monasca,
an additional target_hostname parameter can be configured. It is
the network hostname or IP Address to check instead of hostname.

Add unit tests as well

Change-Id: I33721764e64ef5079b26f78df84c94ed7a1009e7