Add Constraints support
Adding constraints support to clients is slightly more complex than
services as the clients themselves are listed in upper-constraints.txt
which leads to errors that you can't install a specific version and a
constrained version.

This change adds constraints support by also adding a helper script to
edit the constraints to remove monasca-agent. It also uncaps psutil
which is a temporary workaround for monasca-agent until it is uncapped
globally.

Change-Id: Ib56d758e1b6fd6200eace660fc123f147dbdea18