Add router max bandwidth metrics to ovs plugin
This patch will help operators determine if a node that
hosts routers has enough active routers to potentially exceed
the network capacity of the host.  It can also be used to decide
if routers should be moved around/rebalanced.

Three new metrics will be published if the 'publish_router_capacity'
flag is set to true in the ovs.yaml file (off by default):

  vrouter.max_bw_kb (customer metric)
  ovs.vrouter.max_bw_kb (ops metric)
  ovs.vrouter.host_max_bw_kb (ops metric)

This functionality depends on nova flavor bandwith quotas
being configured:

https://wiki.openstack.org/wiki/InstanceResourceQuota#Bandwidth_limits

At some point in the future when neutron supports router QOS,
this functionality should be extended to also honor that.

Change-Id: Ie00260d99bdd1f761dca28d2a5779b906a196564