Add Kubernetes Connector util class
Class to establish a connection to the Kubernetes API if the
container is running within a Kubernetes container.

In Kubernetes access details are passed into the container in how
to connect to the API. This includes host, port, certs. The class
will configure those details for you.

Also add in method to convert memory from string format to bytes

Taken from original review https://review.openstack.org/#/c/391559/

Change-Id: Iea809ea9ece423ef47efffe7c37f3f6eccd81d44