Add prometheus client plugin
The plugin can be configured two ways:

Scrape metrics from a configured set of endpoints that were set up
via prometheus client libraries.

Attempt to autodetect running endpoints we should be monitoring in
a kubernetes environment by either looking at running services or
pods and evaluate their annotations to see if it is set for scraping.

Change-Id: I6d174daa705aff23fbfece9de759a1718a91050f