Add 'agent_service_name' parameter to monasca-setup
At the end of monasca setup run, the agent service is started.
This, however, assumes that the service name is 'monasca-agent'.
As this is not always the case, let's add an extra argument
to control that. Needless to say, the default value is left
untouched.

Change-Id: I66029329dcf7bf014436e64d96c26ed522c268d4
(cherry picked from commit fa3e4daea8912241eefe11032afbcc438b1aebec)