Add Dimensions into White List
Before submitting metrics, we want to check with the white list to
see what are the metrics we want to report and what are the dimension
we want in these metrics. If the key under "metrics" section is not
in the default dimension, the corresponding value specified in white
list will be used to update the dimension. If there is no key value
pairs under "metrics" section, then check and use the key value pairs
under "dimensions" section.

Change-Id: Ic81cc4c393112eb50f700655cefa839e454354db