Fix version comparison on zookeeper plugin
As it is, the comparison is made between a tuple of strings, which
fails for, for example version '3.4.10' >= '3.3.4', as '10' < '4'.

This change converts the tuple of strings into a list of integers,
and the comparison is made between two list of integers.

Change-Id: Iaedb7c0cfa13c918e70a05e6e20de20c05ca764e
(cherry picked from commit 58a8733526a240428e35d933947d0027a00d918e)