Fix libvirt plugin project/tenant name dimension
Enabling the libvirt plugin to add tenant_name as a dimension is
resulting in "Unable to get tenant list from keystone: Unknown
Attribute: tenants" when using keystone v3. This is due the v3
keystone client uses projects.list() instead of tenants.list().

This change fixes that by checking which client version is being
used and making the function call accordingly.

Change-Id: Icfe88b6a1a7ef193fd998db5a72a7a076d23ffe5