Fix slash handling in KubernetesConnector's get_request
The KubernetesConnector can generate URLs that contain '//' when
joining the API prefix with the endpoint. This causes errors in newer
versions of the Kubernetes API and breaks agent startup since it uses
the API to determine the node hostname.

This patch changes KubernetesConnector.get_request()'s behavior to
remove extraneous slashes when necessary.

Change-Id: Icd45b0eb4a9e5535c7c9d95bda209ec186036fab