Fix Prometheus metrics endpoint scanning
Fix bug where it was failing to determine the right metrics
endpoint. I was seeing messages like the below:

Could not get metrics from http://172.17.0.5:r with error ...

The problem was the += of a string to an array which caused the string
to be split into characters. This fixes the issue by using append
instead of +=

Change-Id: I3e2996077e8013819a47aa868f0e8653892aadfe