Add network congestion detection to monasca
To allow monasca detecting network congestion, we add a new agent plugin for monasca-agent
module based on Explicit Congestion Notification (ECN) tos bits. The agent will create a
map between the number of Congestion Encountered(CE) packets/bytes and the source ip address
responsible of the congestion. The source could be a compute or (if activated) a set of VMs
deployed in a compute. First, we ensure that ECN marking is enabled for both edges. Second,
we check ecn.packets and ecn.bytes and calculate ecn.cong.rate associated to each source of
congestion. Last, the agent sends the three metrics and their timestamp to monasca API for
any further processing and keeps the metric counters into a cache file. This piece of
information could be used to trigger a new kind of performance alarm in monasca threshold.

Change-Id: I98405524588582065c7825ca511c489d59826cda
Task: 5750
Story: 2001234