Fix RabbitMQ whitelist documentation error
In monasca_agent/collector/checks_d/rabbitmq.py QUEUE_TYPE is defined
as 'queues'. In conf.d/rabbitmq.yaml.example the whitelist stores the
permitted queues in the list 'queue'. This should be 'queues'. If the
documentation is followed the user specified white list is ignored.
The same looks to be the case for exchanges and nodes.

Story: 2001290
Change-Id: I3e70a888f5c0db27951e619da8bc17ae9eaf9da9