Update k8s utils and prometheus plugin
For k8s we are updating the pod owner
functionaility

Kubernetes 1.6 and newer supports having owner
reference instead if parsing name. This patch
takes advantage of that.

For prometheus we are adding the ability to look
up pod owner based on a pod dimension. This is
key when using kube-state-metrics so we dont
have dead alarms

Also includes a minor fix in k8s plugin for memory

Patch also includes changes to our agent tests as
we now allow () in our new dimensions keys and values
and our tests were not updated for this change yet

Change-Id: I17ea7f42d4b23534221675309c31feeafa75d20c