import zuul job settings from project-config
Manual step: remove Python 3.5 jobs from stable/pike.

This is a mechanically generated patch to complete step 1 of moving
the zuul job settings out of project-config and into each project
repository.

Because there will be a separate patch on each branch, the branch
specifiers for branch-specific jobs have been removed.

Because this patch is generated by a script, there may be some
cosmetic changes to the layout of the YAML file(s) as the contents are
normalized.

See the python3-first goal document for details:
https://governance.openstack.org/tc/goals/stein/python3-first.html

Change-Id: I5398c9bea0903e5b658918fe209bdd01d7ea1e56
Story: #2002586
Task: #24312